+++ CHANGELOG.rst
@@ -1,12 +1,12 @@
 0.1.0 (In Development)
-======================
+----------------------
 
 0.1.0a2 (2017-05-14)
-====================
+--------------------
 
 - Rename ``timeout_middleware_factory`` to ``timeout_middleware``
 
 0.1.0a1 (2017-05-13)
-====================
+--------------------
 
 - Initial release. Implements timeout middleware